+++ packages/domain-clients-react/src/solana/hooks/useSigningSolanaClient.tsx
@@ -5,6 +5,11 @@ import { useMemo } from 'react';
 import { SolanaClusterId } from '@wallet-ui/react';
 import { useWalletUiSigner } from '@wallet-ui/react';
 
+/***
+ * !!! NOTE: There is a bug where useWalletUiSigner() expects an account in early rendering and throws an error
+ * The workaround is to check for an account this useWalletUiAccount(), and conditionally render the component using this hook if the account exists
+ * GH issue: https://github.com/wallet-ui/wallet-ui/issues/270
+ */
 export const useSigningSolanaClient = ({
   clusterId,
 }: {