+++ packages/domain-clients-core/CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Config constructor for solana
 - _Breaking_ require `hide` prop to determine if to show domain in modal. This is a temporary solution
 
+### Fixed
+
+- Decimal math (converting from micro to base units) should not lose precision
+
 ## [0.2.0] 2025-08-19
 
 ### Changed